M2-PR-02: split tests/mod.rs by subsystem into 10 leaf files
Behaviour-preserving. 189 test attributes preserved (105 in new leaf
files + 86 still in mod.rs awaiting M2-PR-03 lifecycle split).

- focus_flow.rs (711 LOC, 32 tests)
- focus_pipeline.rs (193, 6 — including 2 misfiled from the
  "GNOME Extension State Parsing Tests" banner)
- focus_property.rs (279, 1 proptest block / 5 prop tests + 8 strategy fns)
- autostart.rs (58, 3)
- dbus_naming.rs (303, 28 + 1 proptest block)
- control_commands.rs (31, 4)
- kde_script_paths.rs (69, 4 path tests + build_kde_focus_push)
- sni_presentation.rs (484, 17 — plan estimated 14 but actual audit
  shows 17 legitimate SNI tests)
- gnome_ext_state.rs (46, 3 — the rest of the misnamed banner
  relocated to focus_pipeline.rs and stay in mod.rs)
- config_parsing.rs (119, 8)

Residual tests/mod.rs: 2499 LOC, 86 test attrs — all relocate in
M2-PR-03 (lifecycle subtree).

No production source files modified. cargo build clean;
cargo test --bin kanata-switcher 261/0.
