+++ src/daemon/tests/mod.rs
@@ -13,2088 +13,18 @@ use zbus::zvariant::OwnedObjectPath;
 mod common;
 pub(crate) use common::*;
 
-// === Flow Tests ===
-
-#[test]
-fn test_basic_layer_match() {
-    let rules = vec![rule(Some("firefox"), None, Some("browser"))];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("firefox", ""), "default").unwrap();
-    assert_eq!(
-        actions.actions,
-        vec![FocusAction::ChangeLayer("browser".to_string())]
-    );
-}
-
-#[test]
-fn test_no_match_uses_default() {
-    let rules = vec![rule(Some("firefox"), None, Some("browser"))];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("kitty", ""), "default").unwrap();
-    assert_eq!(
-        actions.actions,
-        vec![FocusAction::ChangeLayer("default".to_string())]
-    );
-}
-
-#[test]
-fn test_same_window_no_action() {
-    let rules = vec![rule(Some("firefox"), None, Some("browser"))];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    handler.handle(&win("firefox", "tab1"), "default");
-    let actions = handler.handle(&win("firefox", "tab1"), "default");
-    assert_eq!(actions, None);
-}
-
-#[test]
-fn test_same_rule_different_window_no_action() {
-    let rules = vec![rule(None, None, Some("global"))];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("firefox", "tab1"), "default").unwrap();
-    assert_eq!(
-        actions.actions,
-        vec![FocusAction::ChangeLayer("global".to_string())]
-    );
-
-    let actions = handler.handle(&win("kitty", "tab2"), "default");
-    assert_eq!(actions, None);
-}
-
-#[test]
-fn test_same_rule_different_window_no_action_with_vk_and_raw() {
-    let rules = vec![Rule {
-        class: None,
-        title: None,
-        on_native_terminal: None,
-        layer: Some("global".to_string()),
-        virtual_key: Some("vk_global".to_string()),
-        raw_vk_action: Some(vec![("vk_raw".to_string(), "Tap".to_string())]),
-        fallthrough: false,
-    }];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("firefox", "tab1"), "default").unwrap();
-    assert_eq!(
-        actions.actions,
-        vec![
-            FocusAction::ChangeLayer("global".to_string()),
-            FocusAction::PressVk("vk_global".to_string()),
-            FocusAction::RawVkAction("vk_raw".to_string(), "Tap".to_string()),
-        ]
-    );
-
-    let actions = handler.handle(&win("kitty", "tab2"), "default");
-    assert_eq!(actions, None);
-}
-#[test]
-fn test_title_change_triggers_action() {
-    let rules = vec![
-        rule(Some("kitty"), Some("vim"), Some("vim")),
-        rule(Some("kitty"), None, Some("terminal")),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("kitty", "bash"), "default").unwrap();
-    assert_eq!(get_layers(&actions), vec!["terminal".to_string()]);
-
-    let actions = handler.handle(&win("kitty", "vim"), "default").unwrap();
-    assert_eq!(get_layers(&actions), vec!["vim".to_string()]);
-}
-
-#[test]
-fn test_unfocus_releases_vk_and_switches_to_default() {
-    let rules = vec![Rule {
-        class: Some("firefox".to_string()),
-        title: None,
-        on_native_terminal: None,
-        layer: Some("browser".to_string()),
-        virtual_key: Some("vk_browser".to_string()),
-        raw_vk_action: None,
-        fallthrough: false,
-    }];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    handler.handle(&win("firefox", ""), "default");
-    let actions = handler.handle(&win("", ""), "default").unwrap();
-
-    assert_eq!(
-        actions.actions,
-        vec![
-            FocusAction::ReleaseVk("vk_browser".to_string()),
-            FocusAction::ChangeLayer("default".to_string()),
-        ]
-    );
-    assert_eq!(actions.new_managed_vks, Vec::<String>::new());
-}
-
-#[test]
-fn test_native_terminal_rule_applies_actions() {
-    let rules = vec![rule(Some("kitty"), None, Some("terminal"))];
-    let native_rule = Some(NativeTerminalRule {
-        layer: "tty".to_string(),
-        virtual_key: Some("vk_tty".to_string()),
-        raw_vk_action: vec![("vk_notify".to_string(), "Tap".to_string())],
-    });
-    let mut handler = FocusHandler::new(rules, native_rule, true);
-
-    let actions = handler
-        .handle(
-            &WindowInfo {
-                class: String::new(),
-                title: String::new(),
-                is_native_terminal: true,
-            },
-            "default",
-        )
-        .unwrap();
-
-    assert!(has_action(
-        &actions,
-        &FocusAction::ChangeLayer("tty".to_string())
-    ));
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk_tty".to_string())
-    ));
-    assert!(has_action(
-        &actions,
-        &FocusAction::RawVkAction("vk_notify".to_string(), "Tap".to_string())
-    ));
-}
-
-#[test]
-fn test_autostart_passthrough_args_skip_oneshot() {
-    let matches = Args::command().get_matches_from([
-        "kanata-switcher",
-        "--install-autostart",
-        "-p",
-        "12000",
-        "--quiet-focus",
-        "--no-indicator",
-    ]);
-    let args = Args::from_arg_matches(&matches).unwrap();
-    let exec_args = autostart_passthrough_args(&matches, &args);
-    assert_eq!(
-        exec_args,
-        vec![
-            "-p".to_string(),
-            "12000".to_string(),
-            "--quiet-focus".to_string(),
-            "--no-indicator".to_string()
-        ]
-    );
-}
-
-#[test]
-fn test_autostart_passthrough_args_indicator_focus_only() {
-    let matches = Args::command().get_matches_from([
-        "kanata-switcher",
-        "--install-autostart",
-        "--indicator-focus-only",
-        "false",
-    ]);
-    let args = Args::from_arg_matches(&matches).unwrap();
-    let exec_args = autostart_passthrough_args(&matches, &args);
-    assert_eq!(
-        exec_args,
-        vec!["--indicator-focus-only".to_string(), "false".to_string()]
-    );
-}
-
-#[test]
-fn test_autostart_desktop_content_escapes_exec() {
-    let exec_path = Path::new("/tmp/kanata switcher");
-    let exec_args = vec![
-        "--quiet-focus".to_string(),
-        "-c".to_string(),
-        "/tmp/config%file.json".to_string(),
-    ];
-    let content = build_autostart_desktop_content(exec_path, &exec_args);
-    assert!(content.contains("Type=Application\n"));
-    assert!(content.contains("Name=Kanata Switcher\n"));
-    assert!(content.contains("X-GNOME-Autostart-enabled=true\n"));
-    assert!(content.contains(
-        "Exec=\"/tmp/kanata switcher\" \"--quiet-focus\" \"-c\" \"/tmp/config%%file.json\"\n"
-    ));
-    assert!(content.contains("TryExec=\"/tmp/kanata switcher\"\n"));
-}
-
-#[test]
-fn test_kwin_query_script_path_scopes_to_uuid_and_query_id() {
-    let path = kwin_query_script_path(7);
-    assert!(path.starts_with("/tmp/kanata-switcher-kwin-query-"));
-    let tail = path
-        .strip_prefix("/tmp/kanata-switcher-kwin-query-")
-        .and_then(|tail| tail.strip_suffix(".js"))
-        .expect("failed to parse query script tail");
-    let mut parts = tail.splitn(3, '-');
-    let _uid = parts.next().expect("missing uid");
-    let query_id = parts.next().expect("missing query id");
-    let request_id = parts.next().expect("missing query request id");
-    assert_eq!(query_id, "7");
-    assert!(Uuid::parse_str(request_id).is_ok());
-    assert_ne!(kwin_query_script_path(7), kwin_query_script_path(8));
-}
-
-#[test]
-fn test_kwin_query_probe_script_path_scopes_to_uuid_and_probe_id() {
-    let path = kwin_query_probe_script_path(11);
-    assert!(path.starts_with("/tmp/kanata-switcher-kwin-query-probe-"));
-    let tail = path
-        .strip_prefix("/tmp/kanata-switcher-kwin-query-probe-")
-        .and_then(|tail| tail.strip_suffix(".js"))
-        .expect("failed to parse probe script tail");
-    let mut parts = tail.splitn(3, '-');
-    let _uid = parts.next().expect("missing uid");
-    let probe_id = parts.next().expect("missing probe id");
-    let request_id = parts.next().expect("missing probe request id");
-    assert_eq!(probe_id, "11");
-    assert!(Uuid::parse_str(request_id).is_ok());
-    assert_ne!(
-        kwin_query_probe_script_path(11),
-        kwin_query_probe_script_path(12)
-    );
-}
-
-#[test]
-fn test_kwin_runtime_script_path_scopes_to_uuid() {
-    let path = kwin_runtime_script_path();
-    assert!(path.starts_with("/tmp/kanata-switcher-kwin-"));
-    assert!(path.ends_with(".js"));
-    let request_id = path
-        .strip_prefix("/tmp/kanata-switcher-kwin-")
-        .and_then(|tail| tail.strip_suffix(".js"))
-        .and_then(|tail| tail.splitn(2, '-').nth(1))
-        .expect("failed to parse runtime uuid");
-    assert!(Uuid::parse_str(request_id).is_ok());
-}
-
-#[test]
-fn test_native_terminal_without_rule_uses_default() {
-    let rules = vec![rule(Some("kitty"), None, Some("terminal"))];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler
-        .handle(
-            &WindowInfo {
-                class: String::new(),
-                title: String::new(),
-                is_native_terminal: true,
-            },
-            "default",
-        )
-        .unwrap();
-
-    assert_eq!(
-        actions.actions,
-        vec![FocusAction::ChangeLayer("default".to_string())]
-    );
-}
-
-#[test]
-fn test_derive_default_dbus_suffix_default_host_default_port() {
-    assert_eq!(derive_default_dbus_suffix("127.0.0.1", 10000), "p10000");
-}
-
-#[test]
-fn test_derive_default_dbus_suffix_default_host_non_default_port() {
-    assert_eq!(derive_default_dbus_suffix("127.0.0.1", 22334), "p22334");
-}
-
-#[test]
-fn test_derive_default_dbus_suffix_non_default_host() {
-    assert_eq!(
-        derive_default_dbus_suffix("192.168.1.2", 22334),
-        "h192_168_1_2_p22334"
-    );
-}
-
-#[test]
-fn test_derive_default_dbus_suffix_host_with_dashes_and_uppercase() {
-    assert_eq!(
-        derive_default_dbus_suffix("MyKB-test", 10000),
-        "hMyKB-test_p10000"
-    );
-}
-
-#[test]
-fn test_derive_default_dbus_suffix_ipv6_style_host() {
-    let result = derive_default_dbus_suffix("::1", 10000);
-    assert_eq!(result, "h__1_p10000");
-    assert!(
-        !result.chars().next().unwrap().is_ascii_digit(),
-        "result must not start with a digit"
-    );
-}
-
-#[test]
-fn test_sanitize_dbus_suffix_alphanumeric_unchanged() {
-    assert_eq!(sanitize_dbus_suffix("kinesis").unwrap(), "kinesis");
-    assert_eq!(sanitize_dbus_suffix("AbC123").unwrap(), "AbC123");
-}
-
-#[test]
-fn test_sanitize_dbus_suffix_replaces_special_chars() {
-    assert_eq!(
-        sanitize_dbus_suffix("my.kb:slot 1").unwrap(),
-        "my_kb_slot_1"
-    );
-}
-
-#[test]
-fn test_sanitize_dbus_suffix_keeps_dashes() {
-    assert_eq!(sanitize_dbus_suffix("foo-bar").unwrap(), "foo-bar");
-}
-
-#[test]
-fn test_sanitize_dbus_suffix_leading_digit_gets_underscore() {
-    assert_eq!(sanitize_dbus_suffix("1kb").unwrap(), "_1kb");
-}
-
-#[test]
-fn test_sanitize_dbus_suffix_empty_rejected() {
-    assert!(matches!(
-        sanitize_dbus_suffix(""),
-        Err(DbusSuffixError::Empty)
-    ));
-}
-
-#[test]
-fn test_sanitize_dbus_suffix_too_long_rejected() {
-    let raw = "a".repeat(65);
-    assert!(matches!(
-        sanitize_dbus_suffix(&raw),
-        Err(DbusSuffixError::TooLong { length: 65, limit: 64 })
-    ));
-}
-
-#[test]
-fn test_sanitize_dbus_suffix_at_max_length_accepted() {
-    let raw = "a".repeat(64);
-    let sanitized = sanitize_dbus_suffix(&raw).expect("64-char input must succeed");
-    assert_eq!(sanitized.chars().count(), 64);
-}
-
-#[test]
-fn test_sanitize_dbus_suffix_max_length_digit_start_rejected() {
-    // 64-char input starting with a digit would expand to 65 chars after the
-    // leading-underscore prepend, exceeding the cap. Must be rejected.
-    let raw = "1".repeat(64);
-    let result = sanitize_dbus_suffix(&raw);
-    assert!(
-        matches!(
-            result,
-            Err(DbusSuffixError::TooLong {
-                length: 65,
-                limit: 64,
-            })
-        ),
-        "64-char digit-start input must overflow the cap, got {:?}",
-        result
-    );
-}
-
-proptest! {
-    /// Property: any non-empty ASCII input within the length cap must sanitize
-    /// successfully *unless* it triggers a digit-prepend overflow (input ==
-    /// 64 chars with a leading digit). Successful outputs satisfy the
-    /// `[A-Za-z0-9_-]` rules, do not start with a digit, fit within the cap,
-    /// and round-trip through `effective_dbus_name` → `is_daemon_bus_name`.
-    #[test]
-    fn proptest_sanitize_dbus_suffix_invariants(input in ".{1,64}") {
-        match sanitize_dbus_suffix(&input) {
-            Ok(sanitized) => {
-                prop_assert!(!sanitized.is_empty(), "sanitized output must be non-empty");
-                prop_assert!(
-                    sanitized.chars().count() <= MAX_DBUS_SUFFIX_LEN,
-                    "sanitized length {} exceeds cap {}",
-                    sanitized.chars().count(),
-                    MAX_DBUS_SUFFIX_LEN
-                );
-                let invalid_char = sanitized
-                    .chars()
-                    .find(|c| !(c.is_ascii_alphanumeric() || *c == '_' || *c == '-'));
-                prop_assert!(
-                    invalid_char.is_none(),
-                    "invalid char {:?} in sanitized output {:?}",
-                    invalid_char,
-                    sanitized
-                );
-                let first = sanitized.chars().next().expect("non-empty");
-                prop_assert!(
-                    !first.is_ascii_digit(),
-                    "sanitized output must not start with a digit, got {:?}",
-                    sanitized
-                );
-                prop_assert!(
-                    is_daemon_bus_name(&effective_dbus_name(&sanitized)),
-                    "effective_dbus_name({:?}) must be recognized as a daemon name",
-                    sanitized
-                );
-            }
-            Err(DbusSuffixError::TooLong { length, limit }) => {
-                // The only failure path reachable from a ≤64-char input is
-                // the digit-prepend overflow: input chars().count() == 64 and
-                // first char is a digit (which survives sanitization as a digit
-                // and therefore triggers the underscore prepend).
-                prop_assert_eq!(limit, MAX_DBUS_SUFFIX_LEN);
-                prop_assert_eq!(length, MAX_DBUS_SUFFIX_LEN + 1);
-                prop_assert_eq!(input.chars().count(), MAX_DBUS_SUFFIX_LEN);
-                let first = input.chars().next().expect("non-empty input");
-                prop_assert!(
-                    first.is_ascii_digit(),
-                    "TooLong from a 64-char input is only valid for digit-start; got {:?}",
-                    input
-                );
-            }
-            Err(other) => prop_assert!(
-                false,
-                "unexpected sanitize error {:?} for input {:?}",
-                other,
-                input
-            ),
-        }
-    }
-}
-
-#[test]
-fn test_resolve_dbus_suffix_explicit_value_wins() {
-    let result = resolve_dbus_suffix(Some("kinesis"), "127.0.0.1", 10000).unwrap();
-    assert_eq!(result, "kinesis");
-}
-
-#[test]
-fn test_resolve_dbus_suffix_explicit_sanitization_applied() {
-    let result = resolve_dbus_suffix(Some("kb.1"), "127.0.0.1", 10000).unwrap();
-    assert_eq!(result, "kb_1");
-}
-
-#[test]
-fn test_resolve_dbus_suffix_explicit_empty_errors() {
-    assert!(resolve_dbus_suffix(Some(""), "127.0.0.1", 10000).is_err());
-}
-
-#[test]
-fn test_resolve_dbus_suffix_defaults_to_derived_suffix() {
-    assert_eq!(
-        resolve_dbus_suffix(None, "127.0.0.1", 10000).unwrap(),
-        "p10000"
-    );
-    assert_eq!(
-        resolve_dbus_suffix(None, "127.0.0.1", 22334).unwrap(),
-        "p22334"
-    );
-}
-
-#[test]
-fn test_effective_dbus_name_compose() {
-    assert_eq!(
-        effective_dbus_name("p10000"),
-        "com.github.kanata.Switcher.instances.p10000"
-    );
-    assert_eq!(
-        effective_dbus_name("kinesis"),
-        "com.github.kanata.Switcher.instances.kinesis"
-    );
-}
-
-#[test]
-fn test_effective_dbus_name_round_trips_through_is_daemon_bus_name() {
-    // Structural invariant — if the prefix used by `effective_dbus_name` and
-    // `is_daemon_bus_name` ever drifts apart, this test catches it.
-    for suffix in ["p10000", "kinesis", "h192_168_1_2_p22334", "_1kb", "a-b-c"] {
-        let name = effective_dbus_name(suffix);
-        assert!(
-            is_daemon_bus_name(&name),
-            "{} produced by effective_dbus_name({:?}) must be recognized",
-            name,
-            suffix
-        );
-    }
-}
-
-#[test]
-fn test_is_daemon_bus_name_accepts_instances_subtree() {
-    assert!(is_daemon_bus_name(
-        "com.github.kanata.Switcher.instances.p10000"
-    ));
-    assert!(is_daemon_bus_name(
-        "com.github.kanata.Switcher.instances.kinesis"
-    ));
-}
-
-#[test]
-fn test_is_daemon_bus_name_rejects_extensions_subtree() {
-    assert!(!is_daemon_bus_name(
-        "com.github.kanata.Switcher.extensions.GNOME"
-    ));
-}
-
-#[test]
-fn test_is_daemon_bus_name_rejects_root_and_bare_prefix() {
-    assert!(!is_daemon_bus_name("com.github.kanata.Switcher"));
-    assert!(!is_daemon_bus_name("com.github.kanata.Switcher.instances"));
-}
-
-#[test]
-fn test_is_daemon_bus_name_rejects_unrelated_names() {
-    assert!(!is_daemon_bus_name("org.gnome.Shell"));
-    assert!(!is_daemon_bus_name("org.kde.KWin"));
-    assert!(!is_daemon_bus_name(""));
-}
-
-#[test]
-fn test_autostart_passthrough_args_includes_dbus_suffix() {
-    let matches = Args::command().get_matches_from([
-        "kanata-switcher",
-        "--install-autostart",
-        "--dbus-suffix",
-        "kinesis",
-    ]);
-    let args = Args::from_arg_matches(&matches).unwrap();
-    let exec_args = autostart_passthrough_args(&matches, &args);
-    assert_eq!(
-        exec_args,
-        vec!["--dbus-suffix".to_string(), "kinesis".to_string()]
-    );
-}
-
-#[test]
-fn test_args_dbus_suffix_rejects_empty() {
-    let result = Args::try_parse_from(["kanata-switcher", "--dbus-suffix", ""]);
-    assert!(
-        result.is_err(),
-        "expected --dbus-suffix '' to error, got {:?}",
-        result.ok().map(|args| args.dbus_suffix)
-    );
-}
-
-#[test]
-fn test_args_dbus_suffix_sanitization_applied() {
-    let args = Args::parse_from(["kanata-switcher", "--dbus-suffix", "kb.1"]);
-    assert_eq!(args.dbus_suffix.as_deref(), Some("kb_1"));
-}
-
-#[test]
-fn test_args_dbus_suffix_at_max_length_boundary() {
-    let exactly_max = "a".repeat(MAX_DBUS_SUFFIX_LEN);
-    let args =
-        Args::try_parse_from(["kanata-switcher", "--dbus-suffix", &exactly_max]).expect(
-            "64-char suffix must be accepted by clap value parser",
-        );
-    assert_eq!(args.dbus_suffix.as_deref(), Some(exactly_max.as_str()));
-
-    let overlong = "a".repeat(MAX_DBUS_SUFFIX_LEN + 1);
-    let result = Args::try_parse_from(["kanata-switcher", "--dbus-suffix", &overlong]);
-    assert!(
-        result.is_err(),
-        "65-char suffix must be rejected by clap value parser"
-    );
-}
-
-#[test]
-fn test_build_kde_focus_push_script_targets_instance_name() {
-    let script = build_kde_focus_push_script(
-        "com.github.kanata.Switcher.instances.kinesis",
-        "windowActivated",
-        "activeWindow",
-    );
-    assert!(
-        script.contains("\"com.github.kanata.Switcher.instances.kinesis\""),
-        "KWin script should reference per-instance bus name; got:\n{}",
-        script
-    );
-    assert!(script.contains("\"com.github.kanata.Switcher\""));
-    assert!(script.contains("WindowFocus"));
-    assert!(script.contains("workspace.windowActivated"));
-    assert!(script.contains("workspace.activeWindow"));
-}
-
-#[test]
-fn test_control_command_restart() {
-    let args = Args::parse_from(["kanata-switcher", "--restart"]);
-    assert_eq!(
-        resolve_control_command(&args),
-        Some(ControlCommand::Restart)
-    );
-}
-
-#[test]
-fn test_control_command_pause() {
-    let args = Args::parse_from(["kanata-switcher", "--pause"]);
-    assert_eq!(resolve_control_command(&args), Some(ControlCommand::Pause));
-}
-
-#[test]
-fn test_control_command_unpause() {
-    let args = Args::parse_from(["kanata-switcher", "--unpause"]);
-    assert_eq!(
-        resolve_control_command(&args),
-        Some(ControlCommand::Unpause)
-    );
-}
-
-#[test]
-fn test_control_command_none() {
-    let args = Args::parse_from(["kanata-switcher"]);
-    assert_eq!(resolve_control_command(&args), None);
-}
-
-#[test]
-fn test_sni_format_layer_letter() {
-    assert_eq!(SniIndicator::format_layer_letter("base"), "B");
-    assert_eq!(SniIndicator::format_layer_letter(""), "?");
-    assert_eq!(SniIndicator::format_layer_letter("  "), "?");
-}
-
-#[test]
-fn test_sni_format_virtual_keys() {
-    assert_eq!(SniIndicator::format_virtual_keys(&[]), "");
-    assert_eq!(
-        SniIndicator::format_virtual_keys(&[String::from("vk_media")]),
-        "V"
-    );
-    assert_eq!(
-        SniIndicator::format_virtual_keys(&[String::from("a"), String::from("b")]),
-        "2"
-    );
-    let keys = vec!["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"]
-        .into_iter()
-        .map(String::from)
-        .collect::<Vec<_>>();
-    assert_eq!(SniIndicator::format_virtual_keys(&keys), "9+");
-}
-
-fn sni_buffer_has_layer_pixels(buffer: &[u8]) -> bool {
-    buffer.chunks_exact(4).any(|chunk| {
-        let [alpha, red, green, blue] = chunk else {
-            return false;
-        };
-        *alpha > 0 && *red > 0 && *red == *green && *green == *blue
-    })
-}
-
-fn sni_buffer_has_vk_pixels(buffer: &[u8]) -> bool {
-    buffer.chunks_exact(4).any(|chunk| {
-        let [alpha, red, green, blue] = chunk else {
-            return false;
-        };
-        *alpha > 0 && *red == 0 && *green > 0 && *blue > 0
-    })
-}
-
-#[test]
-fn test_sni_icon_color_layers_and_vks() {
-    let icon = SniIndicator::render_icon("A", "B");
-    assert!(sni_buffer_has_layer_pixels(&icon.data));
-    assert!(sni_buffer_has_vk_pixels(&icon.data));
-}
-
-#[test]
-fn test_sni_icon_color_layer_only() {
-    let icon = SniIndicator::render_icon("A", "");
-    assert!(sni_buffer_has_layer_pixels(&icon.data));
-    assert!(!sni_buffer_has_vk_pixels(&icon.data));
-}
-
-#[derive(Clone, Default)]
-struct MockSniControlCounts {
-    restart: usize,
-    pause: usize,
-    unpause: usize,
-    quit: usize,
-}
-
-#[derive(Clone)]
-struct MockSniControl {
-    counts: Arc<Mutex<MockSniControlCounts>>,
-}
-
-impl MockSniControl {
-    fn new() -> Self {
-        Self {
-            counts: Arc::new(Mutex::new(MockSniControlCounts::default())),
-        }
-    }
-
-    fn counts(&self) -> MockSniControlCounts {
-        self.counts.lock().unwrap().clone()
-    }
-}
-
-impl SniControlOps for MockSniControl {
-    fn restart(&self) {
-        self.counts.lock().unwrap().restart += 1;
-    }
-
-    fn pause(&self) {
-        self.counts.lock().unwrap().pause += 1;
-    }
-
-    fn unpause(&self) {
-        self.counts.lock().unwrap().unpause += 1;
-    }
-
-    fn quit(&self) {
-        self.counts.lock().unwrap().quit += 1;
-    }
-}
-
-#[derive(Default)]
-struct MockDconfState {
-    get_calls: usize,
-    set_calls: Vec<bool>,
-}
-
-struct MockDconfBackend {
-    state: Arc<Mutex<MockDconfState>>,
-    get_results: Arc<Mutex<Vec<Result<bool, String>>>>,
-    set_results: Arc<Mutex<Vec<Result<(), String>>>>,
-}
-
-impl MockDconfBackend {
-    fn new(
-        state: Arc<Mutex<MockDconfState>>,
-        get_results: Vec<Result<bool, String>>,
-        set_results: Vec<Result<(), String>>,
-    ) -> Self {
-        Self {
-            state,
-            get_results: Arc::new(Mutex::new(get_results)),
-            set_results: Arc::new(Mutex::new(set_results)),
-        }
-    }
-
-    fn next_get_result(&self) -> Result<bool, String> {
-        let mut results = self.get_results.lock().unwrap();
-        if results.is_empty() {
-            return Err("no mock get results left".to_string());
-        }
-        results.remove(0)
-    }
-
-    fn next_set_result(&self) -> Result<(), String> {
-        let mut results = self.set_results.lock().unwrap();
-        if results.is_empty() {
-            return Err("no mock set results left".to_string());
-        }
-        results.remove(0)
-    }
-}
-
-impl DconfBackend for MockDconfBackend {
-    fn get_bool(&self, _key: &str) -> Result<bool, String> {
-        let mut state = self.state.lock().unwrap();
-        state.get_calls += 1;
-        self.next_get_result()
-    }
-
-    fn set_bool(&self, _key: &str, value: bool) -> Result<(), String> {
-        let mut state = self.state.lock().unwrap();
-        state.set_calls.push(value);
-        self.next_set_result()
-    }
-}
-
-fn mock_dconf_backend_sequence(
-    get_results: Vec<Result<bool, String>>,
-    set_results: Vec<Result<(), String>>,
-) -> (Box<dyn DconfBackend>, Arc<Mutex<MockDconfState>>) {
-    let state = Arc::new(Mutex::new(MockDconfState::default()));
-    let backend = MockDconfBackend::new(state.clone(), get_results, set_results);
-    (Box::new(backend), state)
-}
-
-fn mock_dconf_backend(
-    get_result: Result<bool, String>,
-    set_result: Result<(), String>,
-) -> (Box<dyn DconfBackend>, Arc<Mutex<MockDconfState>>) {
-    mock_dconf_backend_sequence(vec![get_result], vec![set_result])
-}
-
-#[test]
-fn test_sni_indicator_state_focus_only() {
-    let initial = StatusSnapshot {
-        layer: "base".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    let mut state = SniIndicatorState::new(initial.clone(), SNI_DEFAULT_SHOW_FOCUS_ONLY);
-    assert_eq!(state.display_status().layer, "base");
-
-    let focus_status = StatusSnapshot {
-        layer: "browser".to_string(),
-        virtual_keys: vec!["vk_browser".to_string()],
-        layer_source: LayerSource::Focus,
-    };
-    state.update_status(focus_status.clone());
-    assert_eq!(state.display_status().layer, "browser");
-
-    state.toggle_focus_only();
-    assert_eq!(state.display_status().layer, "browser");
-
-    let external_status = StatusSnapshot {
-        layer: "external".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    state.update_status(external_status.clone());
-    assert_eq!(state.display_status().layer, "external");
-
-    state.toggle_focus_only();
-    assert_eq!(state.display_status().layer, "browser");
-
-    state.set_paused(true);
-    assert_eq!(state.display_status().layer, "external");
-}
-
-#[test]
-fn test_sni_indicator_state_initial_focus_only_false() {
-    let initial = StatusSnapshot {
-        layer: "base".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    let mut state = SniIndicatorState::new(initial.clone(), false);
-
-    let focus_status = StatusSnapshot {
-        layer: "browser".to_string(),
-        virtual_keys: vec!["vk_browser".to_string()],
-        layer_source: LayerSource::Focus,
-    };
-    state.update_status(focus_status);
-
-    let external_status = StatusSnapshot {
-        layer: "external".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    state.update_status(external_status);
-
-    assert_eq!(state.display_status().layer, "external");
-}
-
-#[test]
-fn test_sni_settings_store_reads_from_dconf() {
-    let (backend, state) = mock_dconf_backend(Ok(false), Ok(()));
-    let mut store = SniSettingsStore::with_backend(backend);
-    let value = store.read_focus_only();
-    assert_eq!(value, Some(false));
-    let state = state.lock().unwrap();
-    assert_eq!(state.get_calls, 1);
-    assert!(state.set_calls.is_empty());
-}
-
-#[test]
-fn test_sni_settings_store_writes_to_dconf() {
-    let (backend, state) = mock_dconf_backend(Ok(true), Ok(()));
-    let mut store = SniSettingsStore::with_backend(backend);
-    store.write_focus_only(false);
-    let state = state.lock().unwrap();
-    assert_eq!(state.set_calls, vec![false]);
-}
-
-#[test]
-fn test_sni_settings_store_read_error_disables_write() {
-    let (backend, state) = mock_dconf_backend(Err("No such file or directory".to_string()), Ok(()));
-    let mut store = SniSettingsStore::with_backend(backend);
-    let value = store.read_focus_only();
-    assert_eq!(value, None);
-    store.write_focus_only(true);
-    let state = state.lock().unwrap();
-    assert_eq!(state.get_calls, 1);
-    assert!(state.set_calls.is_empty());
-}
-
-#[test]
-fn test_sni_settings_store_key_not_set_allows_write() {
-    let (backend, state) = mock_dconf_backend(Err("key not set".to_string()), Ok(()));
-    let mut store = SniSettingsStore::with_backend(backend);
-    let value = store.read_focus_only();
-    assert_eq!(value, None);
-    store.write_focus_only(true);
-    let state = state.lock().unwrap();
-    assert_eq!(state.get_calls, 1);
-    assert_eq!(state.set_calls, vec![true]);
-}
-
-#[test]
-fn test_resolve_sni_focus_only_override_skips_dconf() {
-    let (backend, state) = mock_dconf_backend(Ok(false), Ok(()));
-    let mut store = SniSettingsStore::with_backend(backend);
-    let value = resolve_sni_focus_only(Some(TrayFocusOnly::True), &mut store);
-    assert!(value);
-    let state = state.lock().unwrap();
-    assert_eq!(state.get_calls, 0);
-    assert!(state.set_calls.is_empty());
-}
-
-#[test]
-fn test_sni_toggle_persists_to_dconf() {
-    let (backend, state) = mock_dconf_backend(Ok(true), Ok(()));
-    let store = SniSettingsStore::with_backend(backend);
-    let initial = StatusSnapshot {
-        layer: "base".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    let control = MockSniControl::new();
-    let (menu_refresh, _menu_receiver) = MenuRefresh::new();
-    let mut indicator = SniIndicator {
-        state: SniIndicatorState::new(initial, true),
-        control: Arc::new(control),
-        settings: store,
-        menu_refresh,
-    };
-
-    indicator.toggle_focus_only();
-    let state = state.lock().unwrap();
-    assert_eq!(state.set_calls, vec![false]);
-}
-
-#[test]
-fn test_sni_toggle_sends_menu_refresh() {
-    let (menu_refresh, receiver) = MenuRefresh::new();
-    let receiver = receiver;
-    let initial = StatusSnapshot {
-        layer: "base".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    let control = MockSniControl::new();
-    let mut indicator = SniIndicator {
-        state: SniIndicatorState::new(initial, SNI_DEFAULT_SHOW_FOCUS_ONLY),
-        control: Arc::new(control),
-        settings: SniSettingsStore::disabled(),
-        menu_refresh,
-    };
-
-    indicator.toggle_focus_only();
-    assert_eq!(*receiver.borrow(), 1);
-}
-
-#[test]
-fn test_sni_menu_actions_dispatch_control() {
-    let initial = StatusSnapshot {
-        layer: "base".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    let control = MockSniControl::new();
-    let control_counts = control.clone();
-    let (menu_refresh, _menu_receiver) = MenuRefresh::new();
-    let mut indicator = SniIndicator {
-        state: SniIndicatorState::new(initial, SNI_DEFAULT_SHOW_FOCUS_ONLY),
-        control: Arc::new(control),
-        settings: SniSettingsStore::disabled(),
-        menu_refresh,
-    };
-
-    let menu = indicator.menu();
-    let mut found_pause = false;
-    let mut found_restart = false;
-    let mut found_quit = false;
-    for item in menu {
-        match item {
-            MenuItem::Checkmark(check) if check.label == "Pause" => {
-                found_pause = true;
-                (check.activate)(&mut indicator);
-            }
-            MenuItem::Standard(standard) if standard.label == "Restart" => {
-                found_restart = true;
-                (standard.activate)(&mut indicator);
-            }
-            MenuItem::Standard(standard) if standard.label == "Quit" => {
-                found_quit = true;
-                (standard.activate)(&mut indicator);
-            }
-            _ => {}
-        }
-    }
-
-    assert!(found_pause);
-    assert!(found_restart);
-    assert!(found_quit);
-    let counts = control_counts.counts();
-    assert_eq!(counts.pause, 1);
-    assert_eq!(counts.restart, 1);
-    assert_eq!(counts.quit, 1);
-}
-
-#[test]
-fn test_sni_menu_toggle_affects_display() {
-    let initial = StatusSnapshot {
-        layer: "base".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    let control = MockSniControl::new();
-    let (menu_refresh, _menu_receiver) = MenuRefresh::new();
-    let mut indicator = SniIndicator {
-        state: SniIndicatorState::new(initial, SNI_DEFAULT_SHOW_FOCUS_ONLY),
-        control: Arc::new(control),
-        settings: SniSettingsStore::disabled(),
-        menu_refresh,
-    };
-
-    let focus_status = StatusSnapshot {
-        layer: "browser".to_string(),
-        virtual_keys: vec!["vk_browser".to_string()],
-        layer_source: LayerSource::Focus,
-    };
-    indicator.update_status(focus_status);
-
-    let (layer_text, _) = indicator.display_strings();
-    assert_eq!(layer_text, "B");
-
-    let external_status = StatusSnapshot {
-        layer: "external".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    indicator.update_status(external_status);
-
-    indicator.toggle_focus_only();
-    let (layer_text, vk_text) = indicator.display_strings();
-    assert_eq!(layer_text, "E");
-    assert!(vk_text.is_empty());
-
-    indicator.toggle_focus_only();
-    let (layer_text, vk_text) = indicator.display_strings();
-    assert_eq!(layer_text, "B");
-    assert_eq!(vk_text, "V");
-
-    let tooltip = indicator.tooltip_text();
-    assert!(tooltip.contains("Layer:"));
-}
-
-#[test]
-fn test_sni_tooltip_includes_virtual_keys() {
-    let initial = StatusSnapshot {
-        layer: "base".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    let control = MockSniControl::new();
-    let (menu_refresh, _menu_receiver) = MenuRefresh::new();
-    let mut indicator = SniIndicator {
-        state: SniIndicatorState::new(initial, SNI_DEFAULT_SHOW_FOCUS_ONLY),
-        control: Arc::new(control),
-        settings: SniSettingsStore::disabled(),
-        menu_refresh,
-    };
-
-    let focus_status = StatusSnapshot {
-        layer: "browser".to_string(),
-        virtual_keys: vec!["vk_browser".to_string(), "vk_media".to_string()],
-        layer_source: LayerSource::Focus,
-    };
-    indicator.update_status(focus_status);
-    let tooltip = indicator.tooltip_text();
-    assert!(tooltip.contains("Layer: browser"));
-    assert!(tooltip.contains("vk_browser"));
-    assert!(tooltip.contains("vk_media"));
-}
-
-#[test]
-fn test_sni_title_text_is_single_line() {
-    let initial = StatusSnapshot {
-        layer: "base".to_string(),
-        virtual_keys: Vec::new(),
-        layer_source: LayerSource::External,
-    };
-    let control = MockSniControl::new();
-    let (menu_refresh, _menu_receiver) = MenuRefresh::new();
-    let mut indicator = SniIndicator {
-        state: SniIndicatorState::new(initial, SNI_DEFAULT_SHOW_FOCUS_ONLY),
-        control: Arc::new(control),
-        settings: SniSettingsStore::disabled(),
-        menu_refresh,
-    };
-
-    let focus_status = StatusSnapshot {
-        layer: "browser".to_string(),
-        virtual_keys: vec!["vk_browser".to_string(), "vk_media".to_string()],
-        layer_source: LayerSource::Focus,
-    };
-    indicator.update_status(focus_status);
-
-    let title = indicator.title_text();
-    assert!(title == "Kanata Switcher");
-}
-
-#[tokio::test]
-async fn test_update_status_for_focus_updates_snapshot() {
-    let rules = vec![rule(Some("firefox"), None, Some("browser"))];
-    let handler = Arc::new(Mutex::new(FocusHandler::new(rules, None, true)));
-    let status_broadcaster = StatusBroadcaster::new();
-    let kanata = KanataClient::new("127.0.0.1", 10000, None, true, status_broadcaster.clone());
-
-    let win = win("firefox", "");
-    let actions =
-        update_status_for_focus(&handler, &status_broadcaster, &win, &kanata, "default").await;
-    assert!(actions.is_some());
-
-    let snapshot = status_broadcaster.snapshot();
-    assert_eq!(snapshot.layer, "browser");
-    assert_eq!(snapshot.layer_source, LayerSource::Focus);
-}
-
-#[tokio::test]
-async fn test_update_status_for_focus_unknown_layer_uses_default() {
-    let rules = vec![rule(Some("firefox"), None, Some("browser"))];
-    let handler = Arc::new(Mutex::new(FocusHandler::new(rules, None, true)));
-    let status_broadcaster = StatusBroadcaster::new();
-    let kanata = KanataClient::new(
-        "127.0.0.1",
-        10000,
-        Some("default".to_string()),
-        true,
-        status_broadcaster.clone(),
-    );
-
-    {
-        let mut inner = kanata.inner.try_lock().expect("Expected KanataClient lock");
-        inner.known_layers = vec!["default".to_string()];
-    }
-
-    let win = win("firefox", "");
-    let actions =
-        update_status_for_focus(&handler, &status_broadcaster, &win, &kanata, "default").await;
-    assert!(actions.is_some());
-
-    let snapshot = status_broadcaster.snapshot();
-    assert_eq!(snapshot.layer, "default");
-    assert_eq!(snapshot.layer_source, LayerSource::Focus);
-}
-
-#[tokio::test]
-async fn test_handle_focus_event_ignored_when_paused_no_status_change() {
-    let rules = vec![rule(Some("firefox"), None, Some("browser"))];
-    let handler = Arc::new(Mutex::new(FocusHandler::new(rules, None, true)));
-    let status_broadcaster = StatusBroadcaster::new();
-    let pause_broadcaster = PauseBroadcaster::new();
-    let kanata = KanataClient::new("127.0.0.1", 10000, None, true, status_broadcaster.clone());
-
-    pause_broadcaster.set_paused(true);
-    let win = win("firefox", "");
-    let actions = handle_focus_event(
-        &handler,
-        &status_broadcaster,
-        &pause_broadcaster,
-        &win,
-        &kanata,
-        "default",
-    )
-    .await;
-    assert!(actions.is_none());
-    let snapshot = status_broadcaster.snapshot();
-    assert!(snapshot.layer.is_empty());
-
-    pause_broadcaster.set_paused(false);
-    let actions = handle_focus_event(
-        &handler,
-        &status_broadcaster,
-        &pause_broadcaster,
-        &win,
-        &kanata,
-        "default",
-    )
-    .await;
-    assert!(actions.is_some());
-    let snapshot = status_broadcaster.snapshot();
-    assert_eq!(snapshot.layer, "browser");
-    assert_eq!(snapshot.layer_source, LayerSource::Focus);
-}
-
-#[tokio::test]
-async fn test_handle_focus_event_unfocus_paused_does_not_switch_layer() {
-    let rules = vec![rule(Some("firefox"), None, Some("browser"))];
-    let handler = Arc::new(Mutex::new(FocusHandler::new(rules, None, true)));
-    let status_broadcaster = StatusBroadcaster::new();
-    let pause_broadcaster = PauseBroadcaster::new();
-    let kanata = KanataClient::new("127.0.0.1", 10000, None, true, status_broadcaster.clone());
-
-    status_broadcaster.update_layer("current".to_string(), LayerSource::External);
-    pause_broadcaster.set_paused(true);
-
-    let actions = handle_focus_event(
-        &handler,
-        &status_broadcaster,
-        &pause_broadcaster,
-        &WindowInfo::default(),
-        &kanata,
-        "default",
-    )
-    .await;
-    assert!(actions.is_none());
-    let snapshot = status_broadcaster.snapshot();
-    assert_eq!(snapshot.layer, "current");
-    assert_eq!(snapshot.layer_source, LayerSource::External);
-}
-
-#[test]
-fn test_paused_status_resets_virtual_keys_and_source() {
-    let status_broadcaster = StatusBroadcaster::new();
-    status_broadcaster.update_layer("external".to_string(), LayerSource::External);
-    status_broadcaster.update_virtual_keys(vec!["vk_browser".to_string()]);
-    status_broadcaster.set_paused_status("base".to_string());
-    let snapshot = status_broadcaster.snapshot();
-    assert_eq!(snapshot.layer, "base");
-    assert!(snapshot.virtual_keys.is_empty());
-    assert_eq!(snapshot.layer_source, LayerSource::External);
-}
-
-#[test]
-fn test_virtual_key_press_on_focus() {
-    let rules = vec![rule_vk(Some("firefox"), "vk_browser")];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("firefox", ""), "default").unwrap();
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk_browser".to_string())
-    ));
-    assert!(
-        !actions
-            .actions
-            .iter()
-            .any(|a| matches!(a, FocusAction::ReleaseVk(_)))
-    );
-    assert_eq!(actions.new_managed_vks, vec!["vk_browser".to_string()]);
-}
-
-#[test]
-fn test_virtual_key_release_on_switch() {
-    let rules = vec![
-        rule_vk(Some("firefox"), "vk_browser"),
-        rule_vk(Some("kitty"), "vk_terminal"),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    handler.handle(&win("firefox", ""), "default");
-    let actions = handler.handle(&win("kitty", ""), "default").unwrap();
-
-    // Release comes before Press in the action list
-    let release_idx = actions
-        .actions
-        .iter()
-        .position(|a| matches!(a, FocusAction::ReleaseVk(_)));
-    let press_idx = actions
-        .actions
-        .iter()
-        .position(|a| matches!(a, FocusAction::PressVk(_)));
-    assert!(release_idx.unwrap() < press_idx.unwrap());
-
-    assert!(has_action(
-        &actions,
-        &FocusAction::ReleaseVk("vk_browser".to_string())
-    ));
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk_terminal".to_string())
-    ));
-}
-
-#[test]
-fn test_virtual_key_no_change_no_press() {
-    let rules = vec![rule_vk(Some("firefox"), "vk_browser")];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    handler.handle(&win("firefox", "tab1"), "default");
-    let actions = handler.handle(&win("firefox", "tab2"), "default");
-
-    // Window changed but VK is the same - no VK actions (VK already held)
-    assert!(
-        actions.is_none()
-            || !actions
-                .as_ref()
-                .unwrap()
-                .actions
-                .iter()
-                .any(|a| matches!(a, FocusAction::PressVk(_) | FocusAction::ReleaseVk(_)))
-    );
-}
-
-#[test]
-fn test_partial_vk_set_change_only_releases_removed() {
-    // Two rules with fallthrough: vk1 and vk2 are both held
-    // Then switch to a window that only matches vk2 - only vk1 should be released
-    let rules = vec![
-        Rule {
-            class: Some("app".to_string()),
-            title: Some("both".to_string()),
-            on_native_terminal: None,
-            layer: None,
-            virtual_key: Some("vk1".to_string()),
-            raw_vk_action: None,
-            fallthrough: true,
-        },
-        Rule {
-            class: Some("app".to_string()),
-            title: None,
-            on_native_terminal: None,
-            layer: None,
-            virtual_key: Some("vk2".to_string()),
-            raw_vk_action: None,
-            fallthrough: false,
-        },
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    // Focus window that matches both rules - both VKs pressed
-    let actions = handler.handle(&win("app", "both"), "default").unwrap();
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk1".to_string())
-    ));
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk2".to_string())
-    ));
-    assert_eq!(
-        actions.new_managed_vks,
-        vec!["vk1".to_string(), "vk2".to_string()]
-    );
-
-    // Focus window that only matches second rule - only vk1 should be released, vk2 stays held
-    let actions = handler.handle(&win("app", "other"), "default").unwrap();
-    assert!(has_action(
-        &actions,
-        &FocusAction::ReleaseVk("vk1".to_string())
-    ));
-    assert!(!has_action(
-        &actions,
-        &FocusAction::ReleaseVk("vk2".to_string())
-    ));
-    assert!(!has_action(
-        &actions,
-        &FocusAction::PressVk("vk2".to_string())
-    )); // vk2 already held
-    assert_eq!(actions.new_managed_vks, vec!["vk2".to_string()]);
-}
-
-#[test]
-fn test_unfocus_releases_multiple_vks_in_reverse_order() {
-    // Multiple VKs held should be released in reverse order (bottom-to-top)
-    let rules = vec![
-        Rule {
-            class: Some("app".to_string()),
-            title: None,
-            on_native_terminal: None,
-            layer: None,
-            virtual_key: Some("vk1".to_string()),
-            raw_vk_action: None,
-            fallthrough: true,
-        },
-        Rule {
-            class: Some("app".to_string()),
-            title: None,
-            on_native_terminal: None,
-            layer: None,
-            virtual_key: Some("vk2".to_string()),
-            raw_vk_action: None,
-            fallthrough: true,
-        },
-        Rule {
-            class: Some("app".to_string()),
-            title: None,
-            on_native_terminal: None,
-            layer: None,
-            virtual_key: Some("vk3".to_string()),
-            raw_vk_action: None,
-            fallthrough: false,
-        },
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    // Focus window - all three VKs pressed in order
-    let actions = handler.handle(&win("app", ""), "default").unwrap();
-    assert_eq!(
-        actions.new_managed_vks,
-        vec!["vk1".to_string(), "vk2".to_string(), "vk3".to_string()]
-    );
-
-    // Unfocus - all VKs should be released in reverse order (vk3, vk2, vk1)
-    let actions = handler.handle(&win("", ""), "default").unwrap();
-    let release_actions: Vec<_> = actions
-        .actions
-        .iter()
-        .filter_map(|a| {
-            if let FocusAction::ReleaseVk(vk) = a {
-                Some(vk.clone())
-            } else {
-                None
-            }
-        })
-        .collect();
-    assert_eq!(
-        release_actions,
-        vec!["vk3".to_string(), "vk2".to_string(), "vk1".to_string()]
-    );
-}
-
-#[test]
-fn test_raw_vk_action_fires_on_focus() {
-    let rules = vec![rule_raw_vk(
-        Some("firefox"),
-        vec![("vk1", "Tap"), ("vk2", "Toggle")],
-    )];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("firefox", ""), "default").unwrap();
-    assert_eq!(
-        get_raw_vk_actions(&actions),
-        vec![
-            ("vk1".to_string(), "Tap".to_string()),
-            ("vk2".to_string(), "Toggle".to_string()),
-        ]
-    );
-}
-
-#[test]
-fn test_fallthrough_collects_all_layers() {
-    let rules = vec![
-        rule_with_fallthrough(rule(Some("kitty"), None, Some("layer1"))),
-        rule(Some("kitty"), None, Some("layer2")),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("kitty", ""), "default").unwrap();
-    // Both layers should be in the action list, in order
-    assert_eq!(
-        get_layers(&actions),
-        vec!["layer1".to_string(), "layer2".to_string()]
-    );
-}
-
-#[test]
-fn test_fallthrough_add_remove_rule_only_new_actions() {
-    let rules = vec![
-        rule_with_fallthrough(rule(Some("app"), None, Some("base"))),
-        rule(Some("app"), Some("special"), Some("special")),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("app", "special"), "default").unwrap();
-    assert_eq!(
-        get_layers(&actions),
-        vec!["base".to_string(), "special".to_string()]
-    );
-
-    let actions = handler.handle(&win("app", "other"), "default").unwrap();
-    assert_eq!(
-        actions.actions,
-        vec![FocusAction::ChangeLayer("base".to_string())]
-    );
-}
-
-#[test]
-fn test_fallthrough_collects_all_raw_vk_actions() {
-    let rules = vec![
-        rule_with_fallthrough(rule_raw_vk(Some("kitty"), vec![("vk1", "Press")])),
-        rule_raw_vk(Some("kitty"), vec![("vk2", "Tap")]),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("kitty", ""), "default").unwrap();
-    assert_eq!(
-        get_raw_vk_actions(&actions),
-        vec![
-            ("vk1".to_string(), "Press".to_string()),
-            ("vk2".to_string(), "Tap".to_string()),
-        ]
-    );
-}
-
-#[test]
-fn test_fallthrough_all_vks_pressed_and_held() {
-    let rules = vec![
-        rule_with_fallthrough(rule_vk(Some("kitty"), "vk1")),
-        rule_vk(Some("kitty"), "vk2"),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("kitty", ""), "default").unwrap();
-    // Both vk1 and vk2 should be pressed (all matched VKs are held)
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk1".to_string())
-    ));
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk2".to_string())
-    ));
-    assert_eq!(
-        actions.new_managed_vks,
-        vec!["vk1".to_string(), "vk2".to_string()]
-    );
-}
-
-#[test]
-fn test_fallthrough_multiple_vks_all_pressed_and_held() {
-    let rules = vec![
-        rule_with_fallthrough(rule_vk(Some("kitty"), "vk1")),
-        rule_with_fallthrough(rule_vk(Some("kitty"), "vk2")),
-        rule_vk(Some("kitty"), "vk3"),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("kitty", ""), "default").unwrap();
-    // All three VKs should be pressed and held
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk1".to_string())
-    ));
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk2".to_string())
-    ));
-    assert!(has_action(
-        &actions,
-        &FocusAction::PressVk("vk3".to_string())
-    ));
-    assert_eq!(
-        actions.new_managed_vks,
-        vec!["vk1".to_string(), "vk2".to_string(), "vk3".to_string()]
-    );
-}
-
-#[test]
-fn test_fallthrough_action_order_preserved() {
-    // Test that actions from each rule are in order: layer, vk, raw_vk
-    let rules = vec![
-        Rule {
-            class: Some("kitty".to_string()),
-            title: None,
-            on_native_terminal: None,
-            layer: Some("layer1".to_string()),
-            virtual_key: Some("vk1".to_string()),
-            raw_vk_action: Some(vec![("raw1".to_string(), "Tap".to_string())]),
-            fallthrough: true,
-        },
-        Rule {
-            class: Some("kitty".to_string()),
-            title: None,
-            on_native_terminal: None,
-            layer: Some("layer2".to_string()),
-            virtual_key: Some("vk2".to_string()),
-            raw_vk_action: Some(vec![("raw2".to_string(), "Toggle".to_string())]),
-            fallthrough: false,
-        },
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("kitty", ""), "default").unwrap();
-
-    // Expected order: layer1, PressVk(vk1), raw1, layer2, PressVk(vk2), raw2
-    // All matched VKs are pressed (not tapped)
-    assert_eq!(
-        actions.actions,
-        vec![
-            FocusAction::ChangeLayer("layer1".to_string()),
-            FocusAction::PressVk("vk1".to_string()),
-            FocusAction::RawVkAction("raw1".to_string(), "Tap".to_string()),
-            FocusAction::ChangeLayer("layer2".to_string()),
-            FocusAction::PressVk("vk2".to_string()),
-            FocusAction::RawVkAction("raw2".to_string(), "Toggle".to_string()),
-        ]
-    );
-}
-
-#[test]
-fn test_combined_virtual_key_and_raw_vk_action() {
-    let rules = vec![Rule {
-        class: Some("firefox".to_string()),
-        title: None,
-        on_native_terminal: None,
-        layer: Some("browser".to_string()),
-        virtual_key: Some("vk_browser".to_string()),
-        raw_vk_action: Some(vec![("vk_notify".to_string(), "Tap".to_string())]),
-        fallthrough: false,
-    }];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("firefox", ""), "default").unwrap();
-    assert_eq!(
-        actions.actions,
-        vec![
-            FocusAction::ChangeLayer("browser".to_string()),
-            FocusAction::PressVk("vk_browser".to_string()),
-            FocusAction::RawVkAction("vk_notify".to_string(), "Tap".to_string()),
-        ]
-    );
-}
-
-#[test]
-fn test_wildcard_pattern() {
-    let rules = vec![rule(Some("*"), None, Some("any"))];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("anything", ""), "default").unwrap();
-    assert_eq!(get_layers(&actions), vec!["any".to_string()]);
-}
-
-#[test]
-fn test_regex_pattern() {
-    let rules = vec![rule(Some("^(firefox|chrome)$"), None, Some("browser"))];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    assert_eq!(
-        get_layers(&handler.handle(&win("firefox", ""), "default").unwrap()),
-        vec!["browser".to_string()]
-    );
-    assert_eq!(handler.handle(&win("chrome", ""), "default"), None);
-    assert_eq!(
-        get_layers(&handler.handle(&win("chromium", ""), "default").unwrap()),
-        vec!["default".to_string()]
-    );
-}
-
-#[test]
-fn test_three_rules_fallthrough_all_layers_execute() {
-    let rules = vec![
-        rule_with_fallthrough(rule(Some("app"), None, Some("layer1"))),
-        rule_with_fallthrough(rule(Some("app"), None, Some("layer2"))),
-        rule(Some("app"), None, Some("layer3")),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("app", ""), "default").unwrap();
-    assert_eq!(
-        get_layers(&actions),
-        vec![
-            "layer1".to_string(),
-            "layer2".to_string(),
-            "layer3".to_string(),
-        ]
-    );
-}
-
-#[test]
-fn test_multiple_raw_vk_actions_per_rule_all_execute() {
-    let rules = vec![
-        rule_with_fallthrough(rule_raw_vk(
-            Some("app"),
-            vec![("a1", "Press"), ("a2", "Release")],
-        )),
-        rule_raw_vk(Some("app"), vec![("b1", "Tap"), ("b2", "Toggle")]),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("app", ""), "default").unwrap();
-    assert_eq!(
-        get_raw_vk_actions(&actions),
-        vec![
-            ("a1".to_string(), "Press".to_string()),
-            ("a2".to_string(), "Release".to_string()),
-            ("b1".to_string(), "Tap".to_string()),
-            ("b2".to_string(), "Toggle".to_string()),
-        ]
-    );
-}
-
-#[test]
-fn test_non_fallthrough_stops_chain() {
-    // First rule matches but has fallthrough=false, should stop chain
-    let rules = vec![
-        rule(Some("app"), None, Some("layer1")), // fallthrough=false
-        rule(Some("app"), None, Some("layer2")), // would match but shouldn't be reached
-        rule(Some("app"), None, Some("layer3")),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("app", ""), "default").unwrap();
-    // Only layer1 should be collected
-    assert_eq!(get_layers(&actions), vec!["layer1".to_string()]);
-}
-
-#[test]
-fn test_fallthrough_stops_at_non_fallthrough() {
-    // First two rules have fallthrough, third doesn't - chain stops at third
-    let rules = vec![
-        rule_with_fallthrough(rule(Some("app"), None, Some("layer1"))),
-        rule_with_fallthrough(rule(Some("app"), None, Some("layer2"))),
-        rule(Some("app"), None, Some("layer3")), // fallthrough=false, stops here
-        rule(Some("app"), None, Some("layer4")), // should not be reached
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("app", ""), "default").unwrap();
-    assert_eq!(
-        get_layers(&actions),
-        vec![
-            "layer1".to_string(),
-            "layer2".to_string(),
-            "layer3".to_string(),
-        ]
-    );
-    // layer4 should NOT be present
-}
-
-#[test]
-fn test_non_matching_rules_skipped_in_fallthrough() {
-    // Rules that don't match should be skipped even with fallthrough
-    let rules = vec![
-        rule_with_fallthrough(rule(Some("app"), None, Some("layer1"))),
-        rule_with_fallthrough(rule(Some("other"), None, Some("layer2"))), // doesn't match
-        rule(Some("app"), None, Some("layer3")),
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("app", ""), "default").unwrap();
-    // layer2 should be skipped because "other" doesn't match "app"
-    assert_eq!(
-        get_layers(&actions),
-        vec!["layer1".to_string(), "layer3".to_string(),]
-    );
-}
-
-#[test]
-fn test_non_matching_non_fallthrough_rule_does_not_stop_chain() {
-    // A non-matching rule with fallthrough=false should NOT stop the chain
-    // (only matching rules can stop the chain)
-    let rules = vec![
-        rule_with_fallthrough(rule(Some("app"), None, Some("layer1"))),
-        rule(Some("other"), None, Some("layer2")), // doesn't match, fallthrough=false
-        rule(Some("app"), None, Some("layer3")),   // should still be reached
-    ];
-    let mut handler = FocusHandler::new(rules, None, true);
-
-    let actions = handler.handle(&win("app", ""), "default").unwrap();
-    // layer1 and layer3 should be collected; layer2 skipped (doesn't match)
-    // The non-matching rule's fallthrough=false should NOT stop the chain
-    assert_eq!(
-        get_layers(&actions),
-        vec!["layer1".to_string(), "layer3".to_string(),]
-    );
-}
-
-// === Property Tests ===
-
-fn arb_class() -> impl Strategy<Value = String> {
-    prop_oneof![
-        Just("firefox".to_string()),
-        Just("kitty".to_string()),
-        Just("code".to_string()),
-        Just("alacritty".to_string()),
-        Just("".to_string()),
-        "[a-z]{1,10}".prop_map(String::from),
-    ]
-}
-
-fn arb_nonempty_class() -> impl Strategy<Value = String> {
-    arb_class().prop_filter("non-empty class", |class| !class.is_empty())
-}
-
-fn arb_title() -> impl Strategy<Value = String> {
-    prop_oneof![
-        Just("".to_string()),
-        Just("vim".to_string()),
-        Just("bash".to_string()),
-        "[a-zA-Z0-9 ]{0,20}".prop_map(String::from),
-    ]
-}
-
-fn arb_layer() -> impl Strategy<Value = String> {
-    prop_oneof![
-        Just("default".to_string()),
-        Just("browser".to_string()),
-        Just("terminal".to_string()),
-        Just("vim".to_string()),
-        "[a-z]{1,8}".prop_map(String::from),
-    ]
-}
-
-fn arb_vk_name() -> impl Strategy<Value = String> {
-    prop_oneof![
-        Just("vk1".to_string()),
-        Just("vk2".to_string()),
-        Just("vk_browser".to_string()),
-        Just("vk_terminal".to_string()),
-    ]
-}
-
-fn arb_vk_action() -> impl Strategy<Value = String> {
-    prop_oneof![
-        Just("Press".to_string()),
-        Just("Release".to_string()),
-        Just("Tap".to_string()),
-        Just("Toggle".to_string()),
-    ]
-}
-
-fn arb_rule() -> impl Strategy<Value = Rule> {
-    (
-        prop::option::of(arb_class()),
-        prop::option::of(arb_title()),
-        prop::option::of(arb_layer()),
-        prop::option::of(arb_vk_name()),
-        prop::option::of(prop::collection::vec(
-            (arb_vk_name(), arb_vk_action()),
-            0..3,
-        )),
-        any::<bool>(),
-    )
-        .prop_map(|(class, title, layer, vk, raw_vk, fallthrough)| Rule {
-            class,
-            title,
-            on_native_terminal: None,
-            layer,
-            virtual_key: vk,
-            raw_vk_action: raw_vk,
-            fallthrough,
-        })
-}
-
-fn arb_window() -> impl Strategy<Value = WindowInfo> {
-    (arb_class(), arb_title()).prop_map(|(class, title)| WindowInfo {
-        class,
-        title,
-        is_native_terminal: false,
-    })
-}
-
-proptest! {
-    #[test]
-    fn prop_managed_vks_consistent(
-        rules in prop::collection::vec(arb_rule(), 1..5),
-        windows in prop::collection::vec(arb_window(), 1..10),
-    ) {
-        let mut handler = FocusHandler::new(rules, None, true);
-
-        for win in &windows {
-            let _ = handler.handle(win, "default");
-            // Just verify the handler state is consistent (Vec is always valid)
-            prop_assert!(handler.current_virtual_keys.len() <= 10); // sanity bound
-        }
-    }
-
-    #[test]
-    fn prop_releases_before_presses(
-        rules in prop::collection::vec(arb_rule(), 1..5),
-        windows in prop::collection::vec(arb_window(), 2..10),
-    ) {
-        let mut handler = FocusHandler::new(rules, None, true);
-
-        for win in &windows {
-            if let Some(actions) = handler.handle(win, "default") {
-                // All ReleaseVk actions should come before any PressVk actions
-                let first_press_idx = actions.actions.iter().position(|a| matches!(a, FocusAction::PressVk(_)));
-                let last_release_idx = actions.actions.iter().rposition(|a| matches!(a, FocusAction::ReleaseVk(_)));
-
-                if let (Some(press_idx), Some(release_idx)) = (first_press_idx, last_release_idx) {
-                    prop_assert!(release_idx < press_idx, "All releases should come before presses");
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn prop_unfocus_releases_all_vks(
-        rules in prop::collection::vec(arb_rule(), 1..5),
-        win in arb_window(),
-    ) {
-        let mut handler = FocusHandler::new(rules, None, true);
-
-        // Focus a window first
-        let _ = handler.handle(&win, "default");
-        let vks_before = handler.current_virtual_keys.clone();
-
-        // Unfocus (empty class and title)
-        let actions = handler.handle(
-            &WindowInfo {
-                class: String::new(),
-                title: String::new(),
-                is_native_terminal: false,
-            },
-            "default",
-        );
-
-        // All previously active VKs must be released
-        if !vks_before.is_empty() {
-            prop_assert!(actions.is_some());
-            let actions = actions.unwrap();
-            for old_vk in &vks_before {
-                prop_assert!(has_action(&actions, &FocusAction::ReleaseVk(old_vk.clone())));
-            }
-        }
-        // After unfocus, no VKs should be active
-        prop_assert!(handler.current_virtual_keys.is_empty());
-    }
-
-    #[test]
-    fn prop_fallthrough_collects_all_raw_vk(
-        base_class in arb_nonempty_class(),
-        raw_vk1 in prop::collection::vec((arb_vk_name(), arb_vk_action()), 0..2),
-        raw_vk2 in prop::collection::vec((arb_vk_name(), arb_vk_action()), 0..2),
-    ) {
-        let rules = vec![
-            Rule {
-                class: Some(base_class.clone()),
-                title: None,
-                on_native_terminal: None,
-                layer: None,
-                virtual_key: None,
-                raw_vk_action: if raw_vk1.is_empty() { None } else { Some(raw_vk1.clone()) },
-                fallthrough: true,
-            },
-            Rule {
-                class: Some(base_class.clone()),
-                title: None,
-                on_native_terminal: None,
-                layer: None,
-                virtual_key: None,
-                raw_vk_action: if raw_vk2.is_empty() { None } else { Some(raw_vk2.clone()) },
-                fallthrough: false,
-            },
-        ];
-
-        let mut handler = FocusHandler::new(rules, None, true);
-        let win = WindowInfo {
-            class: base_class,
-            title: String::new(),
-            is_native_terminal: false,
-        };
-
-        if let Some(actions) = handler.handle(&win, "default") {
-            let expected: Vec<_> = raw_vk1.into_iter().chain(raw_vk2).collect();
-            prop_assert_eq!(get_raw_vk_actions(&actions), expected);
-        }
-    }
-
-    #[test]
-    fn prop_fallthrough_collects_all_layers(
-        base_class in arb_nonempty_class(),
-        layer1 in arb_layer(),
-        layer2 in arb_layer(),
-    ) {
-        let rules = vec![
-            Rule {
-                class: Some(base_class.clone()),
-                title: None,
-                on_native_terminal: None,
-                layer: Some(layer1.clone()),
-                virtual_key: None,
-                raw_vk_action: None,
-                fallthrough: true,
-            },
-            Rule {
-                class: Some(base_class.clone()),
-                title: None,
-                on_native_terminal: None,
-                layer: Some(layer2.clone()),
-                virtual_key: None,
-                raw_vk_action: None,
-                fallthrough: false,
-            },
-        ];
-
-        let mut handler = FocusHandler::new(rules, None, true);
-        let win = WindowInfo {
-            class: base_class,
-            title: String::new(),
-            is_native_terminal: false,
-        };
-
-        if let Some(actions) = handler.handle(&win, "default") {
-            // Both layers should be collected
-            prop_assert_eq!(get_layers(&actions), vec![layer1, layer2]);
-        }
-    }
-
-    #[test]
-    fn prop_all_matched_vks_pressed_and_held(
-        base_class in arb_nonempty_class(),
-        vk1 in arb_vk_name(),
-        vk2 in arb_vk_name(),
-    ) {
-        let rules = vec![
-            Rule {
-                class: Some(base_class.clone()),
-                title: None,
-                on_native_terminal: None,
-                layer: None,
-                virtual_key: Some(vk1.clone()),
-                raw_vk_action: None,
-                fallthrough: true,
-            },
-            Rule {
-                class: Some(base_class.clone()),
-                title: None,
-                on_native_terminal: None,
-                layer: None,
-                virtual_key: Some(vk2.clone()),
-                raw_vk_action: None,
-                fallthrough: false,
-            },
-        ];
-
-        let mut handler = FocusHandler::new(rules, None, true);
-        let win = WindowInfo {
-            class: base_class,
-            title: String::new(),
-            is_native_terminal: false,
-        };
-
-        if let Some(actions) = handler.handle(&win, "default") {
-            // Both vk1 and vk2 should be pressed (all matched VKs are held)
-            prop_assert!(has_action(&actions, &FocusAction::PressVk(vk1.clone())));
-            prop_assert!(has_action(&actions, &FocusAction::PressVk(vk2.clone())));
-            // new_managed_vks should contain both
-            prop_assert_eq!(actions.new_managed_vks, vec![vk1, vk2]);
-        }
-    }
-}
-
-// === GNOME Extension State Parsing Tests ===
-
-#[test]
-fn test_gnome_extension_state_enabled_f64() {
-    // GNOME Shell returns state as f64, 1.0 = ENABLED
-    use zbus::zvariant::{OwnedValue, Value};
-
-    let mut body = HashMap::new();
-    body.insert(
-        "state".to_string(),
-        OwnedValue::try_from(Value::F64(1.0)).unwrap(),
-    );
-
-    let status = parse_gnome_extension_state(&body);
-    assert!(status.active, "state=1.0 should be active");
-    assert!(status.enabled, "state=1.0 should be enabled");
-    assert!(status.installed, "D-Bus response means installed");
-}
-
-#[test]
-fn test_gnome_extension_state_disabled_f64() {
-    // State 2.0 = DISABLED
-    use zbus::zvariant::{OwnedValue, Value};
-
-    let mut body = HashMap::new();
-    body.insert(
-        "state".to_string(),
-        OwnedValue::try_from(Value::F64(2.0)).unwrap(),
-    );
-
-    let status = parse_gnome_extension_state(&body);
-    assert!(!status.active, "state=2.0 should not be active");
-    assert!(!status.enabled, "state=2.0 should not be enabled");
-}
-
-#[test]
-fn test_gnome_extension_state_missing() {
-    // No state field - should default to not enabled
-    let body = HashMap::new();
-
-    let status = parse_gnome_extension_state(&body);
-    assert!(!status.active, "missing state should not be active");
-    assert!(!status.enabled, "missing state should not be enabled");
-}
+mod focus_flow;
+mod focus_pipeline;
+mod focus_property;
+mod autostart;
+mod dbus_naming;
+mod control_commands;
+mod kde_script_paths;
+mod sni_presentation;
+mod gnome_ext_state;
+mod config_parsing;
+
+// === Misfiled tests (to be moved to lifecycle/ in M2-PR-03) ===
 
 #[tokio::test]
 async fn test_wait_for_restart_or_shutdown_returns_restart() {
@@ -2277,205 +207,6 @@ fn test_decode_logind_object_path_reply_variant() {
     assert_eq!(parsed.as_str(), "/org/freedesktop/login1/session/_1");
 }
 
-#[tokio::test]
-async fn test_update_status_for_focus_filters_invalid_virtual_keys() {
-    // Rule with a virtual key that's NOT in the known list
-    let rules = vec![Rule {
-        class: Some("firefox".to_string()),
-        title: None,
-        on_native_terminal: None,
-        layer: Some("browser".to_string()),
-        virtual_key: Some("invalid_vk".to_string()),
-        raw_vk_action: None,
-        fallthrough: false,
-    }];
-    let handler = Arc::new(Mutex::new(FocusHandler::new(rules, None, true)));
-    let status_broadcaster = StatusBroadcaster::new();
-    let kanata = KanataClient::new(
-        "127.0.0.1",
-        10000,
-        Some("default".to_string()),
-        true,
-        status_broadcaster.clone(),
-    );
-
-    {
-        let mut inner = kanata.inner.try_lock().expect("Expected KanataClient lock");
-        // Set known VKs to a list that does NOT include invalid_vk
-        inner.known_virtual_keys = Some(vec!["valid_vk".to_string()]);
-    }
-
-    let win = win("firefox", "");
-    let _actions =
-        update_status_for_focus(&handler, &status_broadcaster, &win, &kanata, "default").await;
-
-    let snapshot = status_broadcaster.snapshot();
-    // The invalid VK should be filtered out - status should show NO virtual keys
-    assert!(
-        snapshot.virtual_keys.is_empty(),
-        "Invalid VK should not appear in status snapshot, got: {:?}",
-        snapshot.virtual_keys
-    );
-}
-
-#[tokio::test]
-async fn test_update_status_for_focus_shows_valid_virtual_keys() {
-    // Rule with a virtual key that IS in the known list
-    let rules = vec![Rule {
-        class: Some("firefox".to_string()),
-        title: None,
-        on_native_terminal: None,
-        layer: Some("browser".to_string()),
-        virtual_key: Some("vk_browser".to_string()),
-        raw_vk_action: None,
-        fallthrough: false,
-    }];
-    let handler = Arc::new(Mutex::new(FocusHandler::new(rules, None, true)));
-    let status_broadcaster = StatusBroadcaster::new();
-    let kanata = KanataClient::new(
-        "127.0.0.1",
-        10000,
-        Some("default".to_string()),
-        true,
-        status_broadcaster.clone(),
-    );
-
-    {
-        let mut inner = kanata.inner.try_lock().expect("Expected KanataClient lock");
-        // Set known VKs to include vk_browser
-        inner.known_virtual_keys = Some(vec!["vk_browser".to_string()]);
-    }
-
-    let win = win("firefox", "");
-    let _actions =
-        update_status_for_focus(&handler, &status_broadcaster, &win, &kanata, "default").await;
-
-    let snapshot = status_broadcaster.snapshot();
-    // The valid VK should appear in the status
-    assert_eq!(
-        snapshot.virtual_keys,
-        vec!["vk_browser"],
-        "Valid VK should appear in status snapshot"
-    );
-}
-
-// === Config Parsing Tests ===
-
-#[test]
-fn test_config_rejects_unknown_fields() {
-    // "native_terminal" is not a valid field (should be "on_native_terminal")
-    let json = r#"[{"native_terminal": "tty"}]"#;
-    let result: Result<Vec<ConfigEntry>, _> = serde_json::from_str(json);
-    assert!(
-        result.is_err(),
-        "Config should reject unknown field 'native_terminal'"
-    );
-    let err = result.unwrap_err().to_string();
-    assert!(
-        err.contains("unknown field 'native_terminal'"),
-        "Error should mention the unknown field name, got: {}",
-        err
-    );
-}
-
-#[test]
-fn test_config_rejects_typo_in_field_name() {
-    // Common typos should be rejected
-    let json = r#"[{"clas": "firefox", "layer": "browser"}]"#;
-    let result: Result<Vec<ConfigEntry>, _> = serde_json::from_str(json);
-    assert!(result.is_err(), "Config should reject typo 'clas'");
-    let err = result.unwrap_err().to_string();
-    assert!(
-        err.contains("unknown field 'clas'"),
-        "Error should mention the typo field name, got: {}",
-        err
-    );
-}
-
-#[test]
-fn test_config_accepts_valid_rule() {
-    let json = r#"[{"class": "firefox", "layer": "browser"}]"#;
-    let result: Result<Vec<ConfigEntry>, _> = serde_json::from_str(json);
-    assert!(result.is_ok(), "Config should accept valid rule");
-}
-
-#[test]
-fn test_config_accepts_on_native_terminal() {
-    let json = r#"[{"on_native_terminal": "tty"}]"#;
-    let result: Result<Vec<ConfigEntry>, _> = serde_json::from_str(json);
-    assert!(
-        result.is_ok(),
-        "Config should accept 'on_native_terminal' field"
-    );
-}
-
-#[test]
-fn test_config_accepts_default_entry() {
-    let json = r#"[{"default": "base"}]"#;
-    let result: Result<Vec<ConfigEntry>, _> = serde_json::from_str(json);
-    assert!(result.is_ok(), "Config should accept default entry");
-}
-
-#[test]
-fn test_config_parses_matcherless_rule_with_fallthrough() {
-    // A rule with no class/title but with fallthrough: true is valid
-    // (applies to all windows and continues matching)
-    let json = r#"[{"layer": "base", "fallthrough": true}]"#;
-    let result: Result<Vec<ConfigEntry>, _> = serde_json::from_str(json);
-    assert!(
-        result.is_ok(),
-        "Config should parse matcher-less rule with fallthrough"
-    );
-    if let Ok(entries) = result {
-        if let ConfigEntry::Rule(rule) = &entries[0] {
-            assert!(rule.class.is_none());
-            assert!(rule.title.is_none());
-            assert!(rule.fallthrough);
-        } else {
-            panic!("Expected Rule entry");
-        }
-    }
-}
-
-#[test]
-fn test_config_parses_matcherless_rule_without_fallthrough() {
-    // A rule with no class/title and no fallthrough parses but will be rejected
-    // by load_config validation (not tested here since it calls exit)
-    let json = r#"[{"layer": "base"}]"#;
-    let result: Result<Vec<ConfigEntry>, _> = serde_json::from_str(json);
-    assert!(
-        result.is_ok(),
-        "Config should parse (validation happens in load_config)"
-    );
-    if let Ok(entries) = result {
-        if let ConfigEntry::Rule(rule) = &entries[0] {
-            assert!(rule.class.is_none());
-            assert!(rule.title.is_none());
-            assert!(!rule.fallthrough); // default is false
-        } else {
-            panic!("Expected Rule entry");
-        }
-    }
-}
-
-#[test]
-fn test_config_parses_rule_with_class_no_fallthrough() {
-    // A rule with a class matcher doesn't need fallthrough
-    let json = r#"[{"class": "firefox", "layer": "browser"}]"#;
-    let result: Result<Vec<ConfigEntry>, _> = serde_json::from_str(json);
-    assert!(
-        result.is_ok(),
-        "Config should parse rule with class matcher"
-    );
-    if let Ok(entries) = result {
-        if let ConfigEntry::Rule(rule) = &entries[0] {
-            assert!(rule.class.is_some());
-            assert!(!rule.fallthrough);
-        } else {
-            panic!("Expected Rule entry");
-        }
-    }
-}
 
 // === Runtime Lifecycle Tests ===
